change data type of handle to 64 bits for evs_handle_t.
(Logical change 1.127)


git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@447 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/evs.h
@@ -36,7 +36,7 @@
 
 #include <netinet/in.h>
 
-typedef unsigned int evs_handle_t;
+typedef unsigned long long evs_handle_t;
 
 typedef enum {
 	EVS_DISPATCH_ONE,
